Add files via upload
 Project Description:
The Online Voting System is a console-based application developed in C programming language. It simulates the voting process where voters can cast their votes for listed candidates. The system securely counts and displays votes, identifies the winner based on the highest votes, and ensures that each action is reflected in real-time during the program execution.

This project is intended to showcase how basic programming concepts like arrays, loops, conditional statements, and functions can be used to develop a working model of a voting system. It’s a demonstration-level application suitable for academic learning and understanding the core logic behind digital voting mechanisms.